+++ O2GEN/Models/ResourceAllocations.cs
@@ -8,9 +8,9 @@ namespace O2GEN.Models
     /// </summary>
     public class ResourceAllocations
     {
-        public int Id { get; set; } = -1;
-        public int ResourceID { get; set; }
-        public int EngineerID { get; set; }
+        public long Id { get; set; } = -1;
+        public long ResourceID { get; set; }
+        public long EngineerID { get; set; }
         public Guid ObjectUID { get; set; } = Guid.NewGuid();
         public string EngineerName { get; set; }
     }
